requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
itsdangerous>=2.1.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
Flask-Cors>=4.0.0
